Renamed image file for trace to replay.
Removed a lot of hardcoded docs for static tables and enum type structs, so that the diagnostics can properly suggest and warn us when used in the wrong way.

Currently implemented MODERN and COMPACT styles, which are similar and not quite like what the MDI looks like, hence why I did not use MDI as the name for this style.

Will need to implement the drafted MDI style theme on top of this system, now that we have a working baseline.