Adjusted run start/reset detection to hopefully properly re-start a run from the beginning and not carry over leftover data from a previous replay.
Adjusted the UI for bosses, need to test the live side, but the replay works as planned where the timer a boss is defeated is shown on the right side, once the boss is actually defeated. The live side should show the live timer and a difference to show if it's faster or slower.