Added toggle in `/rio settings` to be able to disable the replay frame if the player doesn't want to use it, this way it will unload.
Fixed bug with encounter combat detection not working due to id-mixup after the latest bossID changes.

Updated the combat icon on the top part of the UI and resized the icons, made them a bit smaller.

Added API `RaiderIO.GetCurrentReplay()` that returns the current live and replay summary objects, that contains all the data needed to render the UI, so sharing this in a simple way with others means they can also tap into this data, just beware that it might be volatile as it's not intended as a true public facing API, and neither is it stable if they decide to poke and change the contents as it will break our own addon. If it becomes an issue we might need to consider making deep-copy or preparing unique data to share with the public space to avoid this problem.