Moved the public API block to the bottom as it potentially relies on other modules, but no one should reply on this block.
Made MODERN default style for replays instead of the compact version.

Commented out "MDI" style so it doesn't appear among the choices for style. Might work on it and bring it back at a later date.

Updated the replay public API so that it creates and updates a separate set of tables containing the summary data, and added extra replay data like run url, clear time, dungeon id, enemy forces, short name and name, for extra information. Otherwise it is a table copy so that only external addons will potentially break if user starts editing the contents.