Renamed styles to "Standard" and "Compact"
Added `order` property that we use to help with boss sorting.

Fixed an issue with encounter tracking, moved encounter data in a `encounter` property on the boss.

We still ensure to sort the live bosses based on the replay, but as soon there is a kill we move the live boss on top, and if it doesn't align with the replay we show the route swap icon.

Fixed bug with boss combat detecting for live bosses.