Added `lockReplay` and `replayPoint` properties to the config table.
The replay frame can be locked and unlocked through the cogwheel in the corner.

When locked it will attach itself to the objective frame and act as clickthrough to not interfere with gameplay.

When locked it will become draggable and can be placed anywhere on screen and this position will be remembered.

Clamped the replay timer to the total timer so we don't go over the actual replay total time, this way the timer text is properly colored red when we didn't beat the replay timer.