Moved the special table to `ns.EASTER_EGG` so it is easier to find and edit.
Added a new table `ns.IGNORED_REALMS` to help with flagging certain servers to be ignored.

For example the keystone realm is not flagged as a tournament realm, which makes our addon complain that it's a new server.

`util:IsOnRetailRealm()` will also consider the ignored realms before assuming the server is a live realm.

In `OnPlayerLogin` had to move the line that calls `util:GetRegion()` down so that the realm and realm slug is populated first.