Adding a hotfix for situations where `ENCOUNTER_START` is never called, but `ENCOUNTER_STOP` is...
This can be removed in 10.1.7 when it is fixed, but until then we will simply track that situation by setting `ActiveEncounters` to `false` instead of `nil` when such encounter occurs.

To clean-up we also wipe the table when `PLAYER_ENTERING_WORLD` occurs.

When the boss is defeated, we simply update the `combatStart` to equal the current timer, and bump the `pulls` counter.

Then alter when the boss UI updates we can detect this because `delta` will be `0` so in situations we are showing boss splits and the delta is `0` we just show the kill timer, and skip adding the split that would otherwise just show zero.