Added the `RAID` as type to raids, just to ensure that both dungeons and raids have a type property, that we can rely on to distinguish what a dungeon object represents.
Updated `util:GetLFDStatusForCurrentActivity(activityID, includeExpansionDungeons)` so it now has a boolean to include expansion dungeons, otherwise it will keep to seasonal dungeons, like expected everywhere this function is used. But we have the choice in the future if we ever need the dungeon extracted from the activity, regardless if it's seasonal or not.