Fixed bug in Classic where unit tooltips didn't properly hook.
Reorganized a bit of code to allow showing option to show profile (in the search dialog) as a specific option below the copy URL option.

It only appears if there is a profile to display, unlike the copy URL that always is shown as an option for valid units.

This also affects all dropdowns like the who frame, or other places.