Updated `GetNameRealmForMenu` so that if `accountInfo` is present we use it to extract the character information.
There might be an issue with cross-client where `accountInfo.gameAccountInfo` doesn't contain the `realmName` between client flavors. If this occurs, we simply won't display the dropdown option to copy URL because the realm will be wrong.

The other improvement is the fact we now support `friendsList` if it's present we will extract any missing data from the relevant API.